python_backend/requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
psycopg2-binary
python-multipart
python-dotenv
openpyxl
